API導入　修正
